Moved utility functions into their own files.
