+++ Makefile
@@ -7,7 +7,7 @@ OBJDIR := build
 VPATH = $(shell find src/ -maxdepth 1 -mindepth 1)
 SOURCES_STANDARD := Config.cpp ConfigStatement.cpp HttpContext.cpp Ipv4Address.cpp Lexer.cpp \
 LocationContext.cpp Parser.cpp Port.cpp Scanner.cpp ServerContext.cpp HTTPStatusCode.cpp \
-main.cpp Request.cpp HTTPStatusLine.cpp Response.cpp socket.cpp
+main.cpp Request.cpp HTTPStatusLine.cpp Response.cpp socket.cpp Utilities.cpp
 NAME := webserv
 OBJECTS_STANDARD := $(SOURCES_STANDARD:%.cpp=$(OBJDIR)/%.o)
 OBJECTS_SHARED := $(SOURCES_SHARED:%.cpp=$(OBJDIR)/%.o)
