+++ src/config/Ipv4Address.cpp
@@ -6,7 +6,7 @@
 /*   By: rmeuzela <rmeuzela@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/25 16:47:25 by rmeuzela      #+#    #+#                 */
-/*   Updated: 2025/05/26 20:17:27 by rmeuzela      ########   odam.nl         */
+/*   Updated: 2025/06/04 16:16:14 by rmeuzela      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 #include <limits>
 #include <stdexcept>
 #include "Ipv4Address.hpp"
+#include "Utilities.hpp"
 
 static uint8_t octet_from_string(const std::string& octet)
 {
@@ -27,28 +28,6 @@ static uint8_t octet_from_string(const std::string& octet)
         throw std::runtime_error("Octet out of range.");
     }
     return (stoi_ret);
-}
-
-std::vector<std::string> split(const std::string to_split, const char delimiter)
-{
-    const size_t length = to_split.length();
-    std::vector<std::string> split_str;
-    size_t                      i;
-    size_t                      j;
-    
-    i = 0;
-    j = 0;
-    while (i < length)
-    {
-        if (to_split.at(i) == delimiter)
-        {
-            split_str.push_back(to_split.substr(j, i - j));
-            j = i + 1;            
-        }
-        i++;
-    }
-    split_str.push_back(to_split.substr(j, length - j));
-    return split_str;
 }
 
 uint32_t Ipv4Address::from_string(const std::string& str) const
